+++ src/env/createFileSystem.ts
@@ -1,16 +1,16 @@
 import { FileSystem } from './types';
-import { isNodejs } from './isNodejs';
+// import { isNodejs } from './isNodejs';
 
 export function createFileSystem(fs?: any): FileSystem {
-  let requireFsError = '';
-  if (!fs && isNodejs()) {
+  const requireFsError = '';
+  /*   if (!fs && isNodejs()) {
     try {
       // eslint-disable-next-line global-require
-      fs = require('fs');
+      fs = fs;
     } catch (err) {
       requireFsError = (err as any).toString();
     }
-  }
+  } */
 
   const readFile = fs
     ? (filePath: string) => new Promise((resolve, reject) => { fs.readFile(filePath, (err: any, buffer) => (err ? reject(err) : resolve(buffer))); })